Remove default Laravel structures to help make final review a bit easier
With this commit, I am removing a number of things from the generated application, just to make things a bit easier to review eventually (e.g. Auth is not needed for the demo app, so I have removed it and users). This is not something I would normally do (as with most apps, as they grow, you end up needing most of these things) but my goal here is just to optimize for the team's time and not include boilerplate to distract from the actual code.