Remove unique index from fulltext columns
Mysql apparently does not allow using a unique and a fulltext index together. This is something I'd look into a bit more in a real production use-case and re-evaluate my use of fulltext index columns. As it is, it's okay for demo purposes for these to not have unique constraints guarenteed by the database itself.